requirements.txt
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
coveralls==1.5.0
Django==2.1
django-cors-headers==2.4.0
django-extensions==2.1.2
djangorestframework==3.8.2
docopt==0.6.2
idna==2.7
Pillow==2.2.1
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23